pyproject.toml
[project]
name = "sechubman"
version = "0.1.0"
description = "Development flow as Paleofuturistic Python"
readme = "README.md"
authors = [
    {name = "Paleofuturistic Python", email = "me@here.now"},
]
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.2",
]
classifiers = [
    # License
    "License :: OSI Approved :: Apache Software License",

    # Python versions supported by the package
    "Programming Language :: Python :: 3.12",
]

# click is a transient dependency, but broken beyond version 8.2.1
# so this is needed until the following issue is fixed:
# https://github.com/mkdocs/mkdocs/issues/4032
[dependency-groups]
dev = [
    "click<=8.2.1",
    "mkdocs==1.6.1",
    "mkdocstrings[python]==1.0.0",
    "mypy==1.19.0",
    "pyyaml>=6.0.3",
    "ruff==0.14.7",
]

[build-system]
requires = ["uv_build>=0.8.15,<0.10.0"]
build-backend = "uv_build"

[tool.uv]
exclude-newer = "2025-12-04T01:01:01Z"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    "scratch",
]

[tool.ruff.lint]
select = ["ALL"]
# Recommended ignore list for preventing clashes between ruff lint and ruff format.
ignore = [
    # missing-trailing-comma
    "COM812",
    # indentation-with-invalid-multiple
    "E111",
    # indentation-with-invalid-multiple-comment
    "E114",
    # over-indented
    "E117",
    # line-too-long
    "E501",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = [
    # flake8-annotations
    "ANN",
    # pydocstyle
    "D",
    # flake8-pytest-style
    "PT",
]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.mypy]
files = ["src/"]
disallow_untyped_defs = true
follow_untyped_imports = true
